Fixing Controller Logic and Frontend

+++ client/src/components/admin/admin.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect, useContext } from 'react';
 import { NavLink } from 'react-router-dom';
 import "./admin.css";
+import Swal from 'sweetalert2'
 
 import RemoveRedEyeIcon from '@mui/icons-material/RemoveRedEye';
 import EditIcon from '@mui/icons-material/Edit';
@@ -16,7 +17,7 @@ import BadgeIcon from '@mui/icons-material/Badge';
 import { toast } from 'react-toastify';
 
 import AuthServices from '../../ApiServices/authServices.js';
-
+import _ from 'lodash';
 const Admin = () => {
   const [pageNumber, setPageNumber] = useState(1);
   const [getEmployees, setEmployees] = useState([]);
@@ -28,13 +29,28 @@ const Admin = () => {
   // const [updateUserData, setupdateUserData] = useContext(updateUserDataContext);
   const [deleteUserData, setdeleteUserData] = useContext(deleteUserDataContext);
   const [loginData, setloginData] = useContext(loginContext);
-
   const pages = new Array(numberOfPages).fill(null).map((v, i) => i);
+
+  const [data, setData] = useState({
+    designation: "",
+    team: "",
+    reporting: "",
+    joiningDate: false
+  });
+
+
+  const handleChange = ({ currentTarget: input }) => {
+    setData({ ...data, [input.name]: input.value });
+  };
+
+  const checkboxChange = ({ currentTarget: input }) => {
+    setData({ ...data, [input.name]: !input.value.checked });
+  };
 
   const getEmployeesData = async (e) => {
     try {
       if (loginData.userType === 755) {
-        const { data: res } = await AuthServices.getEmployees(pageNumber);
+        const { data: res } = await AuthServices.getEmployees(pageNumber, data);
         if (res) {
           setEmployees(res.completeEmployee);
           setNumberOfPages(res.numberOfPages)
@@ -54,16 +70,34 @@ const Admin = () => {
 
   useEffect(() => {
     getEmployeesData();
-  }, [pageNumber]);  // // Delete Users in Backend
+  }, [pageNumber, data]);  // // Delete Users in Backend
 
   const deleteEmployeeData = async (id) => {
     try {
-      const { data: res } = await AuthServices.deleteEmployee(id);
-      if (res) {
-        getEmployeesData();
-        setdeleteUserData(deleteUserData);
-        toast.success("Employee Deleted Succesfully")
+      const swalFire = await Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      })
+
+      if (swalFire.isConfirmed) {
+        const { data: res } = await AuthServices.deleteEmployee(id);
+        if (res) {
+          Swal.fire(
+            'Deleted!',
+            'Employee has been deleted.',
+            'success'
+          )
+          getEmployeesData();
+          setdeleteUserData(deleteUserData);
+          toast.success("Employee Deleted Succesfully")
+        }
       }
+
     }
     catch (error) {
       toast.error(error.response.data.msg);
@@ -78,6 +112,37 @@ const Admin = () => {
           <h1>
             <BadgeIcon /> Employees
           </h1>
+
+          {loginData.userType === 755 &&
+
+            <form onSubmit={getEmployeesData}>
+              <div className="input_container">
+                <div className="admin_input">
+                  <label htmlFor="designation">Designation: </label>
+                  <input type="text" value={data.designation} onChange={handleChange}
+
+                    name="designation" />
+                </div>
+                <div className="admin_input">
+                  <label htmlFor="team">Team: </label>
+                  <input type="text" value={data.team} onChange={handleChange}
+
+                    name="team" />
+                </div>
+                <div className="admin_input">
+                  <label htmlFor="reporting">Reporting To: </label>
+                  <input type="text" value={data.reporting} onChange={handleChange}
+
+                    name="reporting" />
+                </div>
+                <div className="admin_input">
+                  <label htmlFor="joiningDate">Sort: </label>
+                  <input type="checkbox" value={data.joiningDate} onChange={checkboxChange}
+                    name="joiningDate" />
+                </div>
+              </div>
+            </form>
+          }
           <table className="table mt-5">
             <thead>
               <tr className="table-dark">
@@ -101,7 +166,7 @@ const Admin = () => {
                     <td>{element.designation}</td>
                     <td>{element.reportingTo}</td>
                     <td>{element.team}</td>
-                    <td>{element.joiningDate}</td>
+                    <td>{new Date(element.joiningDate).toLocaleString('en-GB', { timeZone: 'UTC' })}</td>
                     {loginData.userType === 755 && <td className="d-flex justify-content-between">
                       <NavLink to={`/view/${element._id}`}>
                         {' '}
