+++ client/src/components/meeting/meeting.js
@@ -12,8 +12,10 @@ import {
 } from '../context/contextProvider.js';
 
 import { toast } from 'react-toastify';
-
+import GroupsIcon from '@mui/icons-material/Groups';
 import AuthServices from '../../ApiServices/authServices.js';
+import CancelIcon from '@mui/icons-material/Cancel';
+import Swal from 'sweetalert2'
 
 const Meeting = () => {
     const [pageNumber, setPageNumber] = useState(1);
@@ -56,11 +58,61 @@ const Meeting = () => {
 
     const deleteMeetingData = async (id) => {
         try {
-            const { data: res } = await AuthServices.deleteMeeting(id);
-            if (res) {
-                toast.success("Meeting Deleted Successfully");
-                getMeetingsData();
-                setdeleteMeetData(deleteMeetData);
+            const swalFire = await Swal.fire({
+                title: 'Do you want to delete it?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            })
+
+            if (swalFire.isConfirmed) {
+
+                const { data: res } = await AuthServices.deleteMeeting(id);
+                if (res) {
+                    Swal.fire(
+                        'Deleted!',
+                        'Meeting has been deleted.',
+                        'success'
+                    )
+                    toast.success("Meeting Deleted Successfully");
+                    getMeetingsData();
+                    setdeleteMeetData(deleteMeetData);
+                }
+            }
+        }
+        catch (error) {
+            toast.error(error.response.data.msg);
+        }
+    }
+
+    const abortMeeting = async (id) => {
+        try {
+            const swalFire = await Swal.fire({
+                title: 'Do you want to abort it?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, abort it!'
+            })
+
+            if (swalFire.isConfirmed) {
+
+                const { data: res } = await AuthServices.abortMeeting(id);
+                if (res) {
+                    Swal.fire(
+                        'Aborted!',
+                        'You have aborted the meeting.',
+                        'success'
+                    )
+                    toast.success("Meeting Aborted Successfully");
+                    setdeleteMeetData(deleteMeetData);
+                    getMeetingsData();
+                }
             }
         }
         catch (error) {
@@ -72,7 +124,7 @@ const Meeting = () => {
         <>
             <div className="mt-5">
                 <div className="container">
-                    <h1>Meetings</h1>
+                    <h1> <GroupsIcon /> Meetings</h1>
                     <div className="add-btn mt-2">
                         <NavLink className="btn btn-primary navlink" to="/meetings/create">
                             Create Meetings
@@ -84,6 +136,7 @@ const Meeting = () => {
                                 <th scope="col">ID</th>
                                 <th scope="col">Title</th>
                                 <th scope="col">Created By</th>
+                                <th scope="col">Meeting Room</th>
                                 <th scope="col">Participants</th>
                                 <th scope="col">Meet Timing</th>
                                 <th scope="col">Actions</th>
@@ -97,40 +150,49 @@ const Meeting = () => {
                                         <th scope="row">{id + 1}</th>
                                         <td>{element.meetTitle}</td>
                                         <td>{element.meetCreatedBy}</td>
+                                        <td>{element.meetingRoom}</td>
                                         <td> {element.meetMembers.map((v, id) => {
                                             return (
                                                 <tr key={id + 1}><td>{id + 1}) {v}</td></tr>
                                             )
                                         })}</td>
-                                        <td>{new Date(element.meetStartingTime).toISOString() + " - " + new Date(element.meetEndingTime).toISOString()}</td>
-                                        {(loginData.userType === 755 || loginData.email === element.meetCreatedBy) &&
-                                            <td className="d-flex justify-content-between">
-                                                <NavLink to={`/meetings/edit/${element._id}`}>
-                                                    <button className="btn btn-primary">
-                                                        <EditIcon />
+                                        <td>{new Date(element.meetStartingTime).toLocaleString('en-GB', { timeZone: 'UTC' }) + " - " + new Date(element.meetEndingTime).toLocaleString('en-GB', { timeZone: 'UTC' })}</td>
+                                        <td className="d-flex justify-content-between">
+                                            <button
+                                                className="btn btn-dark"
+                                                onClick={() => abortMeeting(element._id)}
+                                            >
+                                                <CancelIcon />
+                                            </button>
+                                            {(loginData.userType === 755 || loginData.email === element.meetCreatedBy) &&
+                                                <>
+                                                    <NavLink to={`/meetings/edit/${element._id}`}>
+                                                        <button className="btn btn-primary">
+                                                            <EditIcon />
+                                                        </button>
+                                                    </NavLink>
+                                                    <button
+                                                        className="btn btn-danger"
+                                                        onClick={() => deleteMeetingData(element._id)}
+                                                    >
+                                                        <DeleteIcon />
                                                     </button>
-                                                </NavLink>
-                                                <button
-                                                    className="btn btn-danger"
-                                                    onClick={() => deleteMeetingData(element._id)}
-                                                >
-                                                    <DeleteIcon />
-                                                </button>
-                                            </td>
-                                        }
+                                                </>
+                                            }
+                                        </td>
                                     </tr>
                                 );
                             }) : <tr><td>No Meetings</td></tr>}
                         </tbody>
                     </table>
 
-                    {loginData.userType === 755 && <ul className="pagination">
+                    {loginData.userType === 755 && <ul className="pagination justify-content-center">
                         <li className="page-item">
                             <button className="page-link" onClick={() => {
                                 setPageNumber(Math.max(0, pageNumber - 1));
                             }} aria-label="Previous">
                                 <span aria-hidden="true">&laquo;</span>
-                                <span className="sr-only">Previous</span>
+                                <span className="sr-only"></span>
                             </button>
                         </li>
                         {pages.map((pageIndex) => {
@@ -146,7 +208,7 @@ const Meeting = () => {
                             }}
                                 aria-label="Next">
                                 <span aria-hidden="true">&raquo;</span>
-                                <span className="sr-only">Next</span>
+                                <span className="sr-only"></span>
                             </button>
                         </li>
                     </ul>}
